+++ docs/changelog.rst
@@ -4,13 +4,21 @@ Changelog
 
 This page documents the release history of the Saline SDK.
 
+Version 0.2.0 (2025-04-11)
+--------------------------
+
+* Query and faucet bug fix
+* Improved Saline queries
+* Counterparty whitelisting
+* Remove "flow"
+
 Version 0.1.0 (2025-04-03)
 --------------------------
 
 Initial release of the Saline SDK.
 
 Features:
- 
+
 * Basic wallet management
 * Transaction creation and signing
 * RPC client for Saline blockchain interaction